+++ motan-core/src/main/java/com/weibo/api/motan/rpc/DefaultRequest.java
@@ -29,7 +29,7 @@ import java.util.Map;
  * @author fishermen
  * @version V1.0 created at: 2013-5-16
  */
-public class DefaultRequest extends AbstractTraceableRequest implements Serializable {
+public class DefaultRequest implements Request, Traceable, Serializable {
 
     private static final long serialVersionUID = 1168814620391610215L;
 
@@ -42,6 +42,7 @@ public class DefaultRequest extends AbstractTraceableRequest implements Serializ
     private long requestId;
     private byte rpcProtocolVersion = RpcProtocolVersion.VERSION_1.getVersion();
     private int serializeNumber = 0;// default serialization is hession2
+    private TraceableContext traceableContext = new TraceableContext();
 
     @Override
     public String getInterfaceName() {
@@ -80,9 +81,8 @@ public class DefaultRequest extends AbstractTraceableRequest implements Serializ
     }
 
     @Override
-    @SuppressWarnings("unchecked")
     public Map<String, String> getAttachments() {
-        return attachments != null ? attachments : Collections.EMPTY_MAP;
+        return attachments != null ? attachments : Collections.<String, String>emptyMap();
     }
 
     public void setAttachments(Map<String, String> attachments) {
@@ -142,4 +142,7 @@ public class DefaultRequest extends AbstractTraceableRequest implements Serializ
         return serializeNumber;
     }
 
+    public TraceableContext getTraceableContext() {
+        return traceableContext;
+    }
 }