+++ motan-transport-netty/src/main/java/com/weibo/api/motan/transport/netty/NettyEncoder.java
@@ -63,7 +63,11 @@ public class NettyEncoder extends OneToOneEncoder {
         if (message instanceof Response) {
             ((Response) message).setAttachment(MotanConstants.CONTENT_LENGTH, String.valueOf(channelBuffer.readableBytes()));
         }
-        MotanFrameworkUtil.logRequestEvent(getRequestId(message), "after encode rpc " + (message instanceof Request ? "request " : "response ") + this.client.getUrl().getServerPortStr(), System.currentTimeMillis());
+        if (message instanceof Request) {
+            MotanFrameworkUtil.logEvent((Request) message, MotanConstants.TRACE_CENCODE);
+        } else if (message instanceof Response) {
+            MotanFrameworkUtil.logEvent((Response) message, MotanConstants.TRACE_SENCODE);
+        }
         return channelBuffer;
     }